remove ServiceCard and ServiceCard_Test
