+++ Sources/CardPayments/Models/Vault.swift
@@ -1,14 +1,14 @@
 import Foundation
 
-/// This class is used to configure a CardRequest for vaulting
+/// Used to configure Vault request
 public struct Vault {
     
-    /// Property used to associate  a payment method with vault of customer with this ID
+    /// Optional. When set, the value of this property will be used to associate a payment method with the vault of a customer with this ID. If not specified a customer ID will be generated.
     public let customerID: String?
 
     /// Creates an instance of a vault object
     /// - Parameters:
-    ///   - customerID: The identifier for customer. If not specified, PayPal will create one
+    /// - customerID: Optional. When set, the value of this property will be used to associate a payment method with the vault of a customer with this ID. If not specified a customer ID will be generated.
     public init(customerID: String? = nil) {
         self.customerID = customerID
     }
